Rename core/models.py to core/data_schemas.py
'models' is ambiguous in an ML codebase (reads as ML models, but these
are data schemas). Rename clarifies that ParseResult/BookEntry are data
shapes, pairing naturally with interfaces.py (behavior contracts).
